backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic-settings==2.7.1
pypdf==5.1.0
python-docx==1.1.2
